requirements.txt
langgraph>=0.2.0
langchain-openai>=0.2.0
python-telegram-bot>=21.0
httpx>=0.27.0
python-dotenv>=1.0.0
